Move PatternSelection to correct package
