+++ app/src/test/java/rectangledbmi/com/pittsburghrealtimetracker/patterns/PatternViewModelTest.java
@@ -8,7 +8,6 @@ import java.util.List;
 
 import rectangledbmi.com.pittsburghrealtimetracker.mock.PatApiMock;
 import rectangledbmi.com.pittsburghrealtimetracker.model.PatApiService;
-import rectangledbmi.com.pittsburghrealtimetracker.patterns.polylines.PatternSelection;
 import rectangledbmi.com.pittsburghrealtimetracker.selection.Route;
 import rx.Subscription;
 import rx.observers.TestSubscriber;
